requirements.txt
APScheduler==3.10.4
PyYAML==6.0.1
mysql-connector-python==8.3.0
numpy==1.26.4
pandas==2.2.1
pymongo==4.6.3
neo4j==5.18.0
certifi==2024.2.2
indexed-priority-queue==0.1.1
